+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 
 * New types for MessagePack extensions compatible with go-option (#459).
 * Added `box.MustNew` wrapper for `box.New` without an error (#448).
+* Method `Release` for `Future` and `Response` interface: 
+  `Get` and `GetTyped` are calling `Release` by defer.
+  `GetResult` and `GetTypedResult` just return result without `Release` (#493).
 
 ### Changed